+++ libs/components/lookup/src/lib/modules/country-field/country-field.component.ts
@@ -96,7 +96,8 @@ export class SkyCountryFieldComponent
   }
 
   /**
-   * Whether to disable the country field.
+   * Whether to disable the country field on template-driven forms. Don't use this input on reactive forms because they may overwrite the input or leave the control out of sync.
+   * To set the disabled state on reactive forms, use the `FormControl` instead.
    * @default false
    */
   @Input()
@@ -108,6 +109,12 @@ export class SkyCountryFieldComponent
     }
 
     this.#_disabled = isDisabled ?? false;
+
+    if (isDisabled) {
+      this.countrySearchFormControl.disable();
+    } else {
+      this.countrySearchFormControl.enable();
+    }
 
     this.#changeDetector.markForCheck();
   }